Relocate upload image hook to api folder; create hook for login; fix auth service
