+++ apps/api/src/auth/auth.service.ts
@@ -30,7 +30,7 @@ export class AuthService {
       const payload = {
         id: user.id,
         email: user.email,
-        role: user.email,
+        role: user.role,
         isSuspended: user.isSuspended,
       };
